examples/v0_2_api/06_manual_filters_and_orders/pyproject.toml
[project]
name = "strawberry-chemist-example-manual-filters-orders"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.1.0",
    "sqlalchemy>=2.0.41",
    "strawberry-chemist==0.2.0",
    "strawberry-graphql==0.211.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
